+++ crates/cel-cxx-ffi/src/runtime.cc
@@ -5,9 +5,9 @@ namespace rust::cel_cxx {
 
 absl::StatusOr<bool> AnyFfiActivationWrapper::FindVariable(
     absl::string_view name,
-    const google::protobuf::DescriptorPool* ABSL_NONNULL descriptor_pool,
-    google::protobuf::MessageFactory* ABSL_NONNULL message_factory,
-    google::protobuf::Arena* ABSL_NONNULL arena, Value* ABSL_NONNULL result) const
+    const google::protobuf::DescriptorPool* absl_nonnull descriptor_pool,
+    google::protobuf::MessageFactory* absl_nonnull message_factory,
+    google::protobuf::Arena* absl_nonnull arena, Value* absl_nonnull result) const
 {
     std::unique_ptr<Value> result_ptr;
     auto status = ffi_->FindVariable(
@@ -39,9 +39,9 @@ std::vector<FunctionOverloadReference> AnyFfiActivationWrapper::FindFunctionOver
 
 absl::StatusOr<Value> AnyFfiFunctionWrapper::Invoke(
     absl::Span<const Value> args,
-    const google::protobuf::DescriptorPool* ABSL_NONNULL descriptor_pool,
-    google::protobuf::MessageFactory* ABSL_NONNULL message_factory,
-    google::protobuf::Arena* ABSL_NONNULL arena,
+    const google::protobuf::DescriptorPool* absl_nonnull descriptor_pool,
+    google::protobuf::MessageFactory* absl_nonnull message_factory,
+    google::protobuf::Arena* absl_nonnull arena,
     absl::Span<const std::string> overload_id) const
 {
     Value result;