+++ crates/cel-cxx-macros/src/lib.rs
@@ -6,15 +6,29 @@ mod symbol;
 
 /// Derive macro for creating opaque types for cxx-cel.
 ///
+/// # Attributes
+/// - `type = "..."`: The type name of the opaque type.
+/// - `display`: Generates `impl std::fmt::Display` with `Debug` trait.
+/// - `display = EXPR`: Generates `impl std::fmt::Display` with the given expression.
+/// - `crate = ...`: The crate name of the cel_cxx crate, if you are using a different crate name in Cargo.toml.
+///
 /// # Example
 ///
 /// ```ignore
 /// #[derive(Opaque, Debug, Clone, PartialEq)]
-/// #[cel_cxx(type = "my_app.my_module.MyType")]
-/// struct MyType {
-///     value: i32,
+/// #[cel_cxx(type = "my_app.my_module.User")]
+/// // Generates `std::fmt::Display` impl with `Debug` trait.
+/// #[cel_cxx(display)]
+/// // or you can specify a custom format.
+/// // Generates `std::fmt::Display` impl with custom format.
+/// #[cel_cxx(display = write!(fmt, "User(id={id}, name={name})", id = self.id, name = self.name))]
+/// struct User {
+///     id: i32,
+///     name: String,
+///     email: String,
 /// }
 /// ```
+/// 
 ///
 #[proc_macro_derive(Opaque, attributes(cel_cxx))]
 pub fn derive_opaque(input: TokenStream) -> TokenStream {