+++ crates/cel-cxx-macros/src/symbol.rs
@@ -1,3 +1,4 @@
 pub(crate) const CEL_CXX: &str = "cel_cxx";
 pub(crate) const TYPE: &str = "type";
 pub(crate) const CRATE: &str = "crate";
+pub(crate) const DISPLAY: &str = "display";